PSFR-1339: removed pathway code that does not use yaml config
